Only use strings needed for tests + move files to test folder
