+++ src/tests/CalendarTooltipItem.test.js
@@ -1,6 +1,6 @@
 import { shallow } from '@vue/test-utils'
 import CalendarTooltipItem from '../components/daysoff/CalendarTooltipItem'
-import i18n_nl from '../i18n/nl.json'
+import i18n_nl from './i18n/nl.json'
 
 const mocks = {
   $t: v => v,
